Extract remote deck source helpers
